Bu Issue: the CoreNetwork input_size was wrong
The input_size of the CoreNetwork passed through from the model.py was `hidden_size`, which is not actually the real input. The real input is loc_hidden+glimpse_hidden, and I changed it to that.

Nobody saw it until now because 128+128=256, which is the same size of the hidden_size of h_t.