Fixed bugs of hidden_size and input parameters in trainer and model
The input_size of the CoreNetwork passed through from the model.py was `hidden_size`, which is not actually the real input. The real input is loc_hidden+glimpse_hidden, and I changed it to that.
Nobody saw it until now because 128+128=256, which is the same size of the hidden_size of h_t.

Moreover, when declaring RecurrentAttention(..), the parameters are swapped. Nobody saw it because glimpse_hidden and locatin_hidden were always the same size (128x128). Although,  if you change size, you can see it.